requirements.txt
asgiref              3.5.0
beautifulsoup4       4.11.1
bs4                  0.0.1
certifi              2022.5.18.1
charset-normalizer   2.0.12
Django               4.0.4
django-cleanup       6.0.0
django-debug-toolbar 3.2.4
django-filter        21.1
django-js-asset      2.0.0
django-light         0.1.0.post3
django-mptt          0.13.4
idna                 3.3
lxml                 4.9.0
Pillow               9.1.0
pip                  22.0.4
python-slugify       6.1.1
requests             2.27.1
setuptools           58.1.0
six                  1.16.0
soupsieve            2.3.2.post1
sqlparse             0.4.2
text-unidecode       1.3
tzdata               2022.1
urllib3              1.26.9